# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "The purpose of this application is to provide some\nuse cases describing how to generate docs for your API",
    "title": "go-swagger examples.",
    "version": "0.0.1"
  },
  "host": "localhost",
  "basePath": "/",
  "paths": {
    "/orders": {
      "get": {
        "description": "List of orders made by current user",
        "tags": [
          "orders"
        ],
        "summary": "Handler returning list of orders.",
        "operationId": "listOrder",
        "parameters": [
          {
            "maximum": 12,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "x-go-name": "Limit",
            "description": "List order limitations.",
            "name": "limit",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "order",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/order"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "Handler to create an order.",
        "operationId": "createOrder",
        "parameters": [
          {
            "x-go-name": "Order",
            "description": "The order to submit.",
            "name": "order",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/order"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/orderResponse"
          },
          "422": {
            "$ref": "#/responses/validationError"
          }
        }
      }
    }
  },
  "definitions": {
    "order": {
      "type": "object",
      "title": "An Order for one or more products by a user.",
      "required": [
        "id",
        "user_id"
      ],
      "properties": {
        "id": {
          "description": "ID of the order",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "items": {
          "description": "items for this order\nmininum items: 1",
          "type": "array",
          "items": {
            "type": "object",
            "required": [
              "product_id",
              "qty"
            ],
            "properties": {
              "product_id": {
                "description": "the id of the product to order",
                "type": "integer",
                "format": "int64",
                "x-go-name": "ProductID"
              },
              "qty": {
                "description": "the quantity of this product to order",
                "type": "integer",
                "format": "int32",
                "minimum": 1,
                "x-go-name": "Quantity"
              }
            }
          },
          "x-go-name": "OrderItems"
        },
        "user_id": {
          "description": "the id of the user who placed the order.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "UserID"
        }
      },
      "x-go-name": "Order",
      "x-go-package": "github.com/valerykalashnikov/go-swagger-examples"
    }
  },
  "responses": {
    "orderResponse": {
      "description": "An OrderResponse response model\n\nThis is used for returning a response with a single order as body",
      "schema": {
        "$ref": "#/definitions/order"
      },
      "headers": {
        "Rate-Limit-Remaining": {
          "type": "string",
          "description": "Get number of rate limted requests remaining"
        }
      }
    },
    "validationError": {
      "description": "A ValidationError is a swagger response to represent error",
      "schema": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "x-go-name": "Code"
          },
          "field": {
            "type": "string",
            "x-go-name": "Field"
          },
          "message": {
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    }
  }
}